requirements.txt
streamlit
openai
slither
python-dotenv